requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.10.0
certifi==2022.9.24
click==8.1.3
Deprecated==1.2.13
Django==4.1.2
django-redis==5.2.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
sqlparse==0.4.3
tomli==2.0.1
wrapt==1.14.1
